+++ src/main/java/ru/practicum/shareit/booking/repository/BookingRepository.java
@@ -8,29 +8,29 @@ import ru.practicum.shareit.booking.model.BookingStatus;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface BookingRepository extends JpaRepository<Booking, Long> {
-    List<Booking> findAllByUserIdOrderByStartDesc(long userId);
+public interface BookingRepository extends JpaRepository<Booking, Integer> {
+    List<Booking> findAllByUserIdOrderByStartDesc(int userId);
 
-    List<Booking> findAllByUserIdAndEndBeforeOrderByStartDesc(long userId, LocalDateTime endDateTime);
+    List<Booking> findAllByUserIdAndEndBeforeOrderByStartDesc(int userId, LocalDateTime endDateTime);
 
-    List<Booking> findAllByUserIdAndStartAfterOrderByStartDesc(long userId, LocalDateTime startDateTime);
+    List<Booking> findAllByUserIdAndStartAfterOrderByStartDesc(int userId, LocalDateTime startDateTime);
 
-    List<Booking> findAllByUserIdAndStartBeforeAndEndAfterOrderByStartDesc(long userId, LocalDateTime startDateTime, LocalDateTime endDateTime);
+    List<Booking> findAllByUserIdAndStartBeforeAndEndAfterOrderByStartDesc(int userId, LocalDateTime startDateTime, LocalDateTime endDateTime);
 
-    List<Booking> findAllByUserIdAndStatusOrderByStartDesc(long userId, BookingStatus bookingStatus);
+    List<Booking> findAllByUserIdAndStatusOrderByStartDesc(int userId, BookingStatus bookingStatus);
 
-    List<Booking> findAllByItemOwnerIdOrderByStartDesc(long ownerId);
+    List<Booking> findAllByItemOwnerIdOrderByStartDesc(int ownerId);
 
-    List<Booking> findAllByItemOwnerIdAndEndBeforeOrderByStartDesc(long ownerId, LocalDateTime endDateTime);
+    List<Booking> findAllByItemOwnerIdAndEndBeforeOrderByStartDesc(int ownerId, LocalDateTime endDateTime);
 
-    List<Booking> findAllByItemOwnerIdAndStartAfterOrderByStartDesc(long ownerId, LocalDateTime startDateTime);
+    List<Booking> findAllByItemOwnerIdAndStartAfterOrderByStartDesc(int ownerId, LocalDateTime startDateTime);
 
-    List<Booking> findAllByItemOwnerIdAndStartBeforeAndEndAfterOrderByStartDesc(long ownerId, LocalDateTime startDateTime, LocalDateTime endDateTime);
+    List<Booking> findAllByItemOwnerIdAndStartBeforeAndEndAfterOrderByStartDesc(int ownerId, LocalDateTime startDateTime, LocalDateTime endDateTime);
 
-    List<Booking> findAllByItemOwnerIdAndStatusOrderByStartDesc(long ownerId, BookingStatus status);
+    List<Booking> findAllByItemOwnerIdAndStatusOrderByStartDesc(int ownerId, BookingStatus status);
 
-    List<Booking> findAllByItemId(long itemId);
+    List<Booking> findAllByItemId(int itemId);
 
     @Query("SELECT b FROM Booking b WHERE b.item.id = :itemId AND b.user.id = :userId AND b.status = ru.practicum.shareit.booking.model.BookingStatus.APPROVED AND b.end < :currentTime")
-    List<Booking> findAllApprovedByItemIdAndUserId(long itemId, long userId, LocalDateTime currentTime);
+    List<Booking> findAllApprovedByItemIdAndUserId(int itemId, int userId, LocalDateTime currentTime);
 }
