+++ src/main/java/ru/practicum/shareit/item/repository/ItemRepository.java
@@ -6,8 +6,8 @@ import ru.practicum.shareit.item.model.Item;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
-    List<Item> findAllByOwnerId(Long ownerId);
+public interface ItemRepository extends JpaRepository<Item, Integer> {
+    List<Item> findAllByOwnerId(int ownerId);
 
     @Query("select i from Item i " +
             "where upper(i.name) like upper(concat('%', ?1, '%')) " +
